Реализована фича "Отзывы"

+++ src/main/java/ru/yandex/practicum/filmorate/mappers/ReviewMap/ReviewExtractor.java
@@ -20,8 +20,8 @@ public class ReviewExtractor implements ResultSetExtractor<Review> {
                 .reviewId(rs.getLong("review_id"))
                 .content(rs.getString("content"))
                 .isPositive(rs.getBoolean("is_positive"))
-                .userId(rs.getLong("user_id"))
-                .filmId(rs.getLong("film_id"))
+                .userId(rs.getInt("user_id"))
+                .filmId(rs.getInt("film_id"))
                 .useful(rs.getInt("useful"))
                 .userReactions(new HashMap<>())
                 .build();
